exemple/pollsapi/requirements.txt
anyjson==0.3.3
Django==2.2.6
django-admin-generator==2.2.0
django-seed==0.1.9
django-utils2==2.5.4
djangorestframework==3.10.3
Faker==2.0.2
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
